Expose more renderscript native type
Currently, Renderscript only expose a subset of its
datatype within the Android system. For instance, the unsigned
versions of the vector types for bytes, int, long and short are
not defined. In addition, while the signed scalar types for
renderscript have an elementary type equivalent, there is no such
equivalence for unsigned values.

This commit defines these missing classes and defines operations
for these types.

Change-Id: I45353a3b3f304051c072472284e7aefac60d127d